Ajout commentaires, changement en types standards

+++ TShirt/analog.h
@@ -1,4 +1,4 @@
 /** Definitions for the analog converter **/
 
 void ad_init(unsigned char channel);
-unsigned int ad_sample(void);
+uint8_t ad_sample(void);
